requirements.txt
discord.py
aiosqlite
python-dotenv
quart
httpx
aiosmtplib